6 done

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ long int c;
 long long int d;
 char f;
 
-printf("Size of char: %ld byte(s)\n", sizeof(f));
-printf("Size of int: %ld byte(s)\n", sizeof(a));
-printf("Size of long int: %ld byte(s)\n", sizeof(c));
-printf("Size of long long int: %ld byte(s)\n", sizeof(d));
-printf("Size of float: %ld byte(s)\n", sizeof(b));
+printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(b));
 return(0);
 }
